dp3/requirements.txt
requests==2.32.3
Pillow==11.2.1
PyYAML==6.0.2
